little code reformat
